backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv

numpy
pandas
scikit-learn
scipy
joblib